refactor(schema-editor): modularize field creation and inner field handling
Refactors bucket field add logic and inner field management into a more modular structure.
Moves tab and state handling for inner fields into a dedicated component, improving separation
of concerns. Centralizes input property computation and simplifies default value logic for
better readability and flexibility. Enhances maintainability and prepares for future feature
expansion.
